final-pjt-back/requirements.txt
asgiref==3.4.1
backcall==0.2.0
colorama==0.4.4
decorator==5.1.0
Django==3.2.9
django-appconf==1.0.4
django-cors-headers==3.7.0
django-imagekit==4.0.2
django-extensions==3.1.3
django-seed==0.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Faker==9.3.1
ipython==7.28.0
jedi==0.18.0
matplotlib-inline==0.1.3
parso==0.8.2
pickleshare==0.7.5
pilkit==2.0
Pillow==8.3.2
prompt-toolkit==3.0.20
psycopg2==2.9.1
Pygments==2.10.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
traitlets==5.1.0
wcwidth==0.2.5